requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
google-auth>=2.27
PyJWT>=2.8
cryptography>=42.0
gunicorn>=21.0
celery>=5.3
redis>=5.0
requests>=2.31.0
google-genai
